Refactoring for numbers UI tests.
List of changes:
       - wait until configuration page is loaded
       - add page object assert for text values in elements
       - move configuration page object to separate file
       - move service details to separate file
       - use page object method instead selenium direct calls
