+++ tests/ui_tests/app/pages.py
@@ -11,9 +11,6 @@
 # limitations under the License.
 
 # Created by a1wen at 05.03.19
-import json
-
-from retrying import retry
 from selenium.webdriver.common.keys import Keys
 from selenium.common.exceptions import TimeoutException, InvalidElementStateException,\
     ElementClickInterceptedException, NoSuchElementException, StaleElementReferenceException
@@ -21,16 +18,10 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait as WDW
 from selenium.webdriver.common.action_chains import ActionChains
-from tests.ui_tests.app.locators import Menu, Common, Cluster, Provider, Host,\
-    ConfigurationLocators, Service
+from tests.ui_tests.app.locators import Menu, Common, Cluster, Provider, Host, Service
 from tests.ui_tests.app.helpers import bys
 from cm.errors import ERRORS
 from time import sleep, time
-
-
-def retry_on_exception(exc):
-    return any((isinstance(exc, StaleElementReferenceException),
-                isinstance(exc, NoSuchElementException)))
 
 
 def repeat_dec(timeout=10, interval=0.1):
@@ -90,7 +81,16 @@ class BasePage:
         return element.find_element(Common.mat_error).text
 
     def _wait_element_present(self, locator: tuple, timer=5):
-        return WDW(self.driver, timer).until(EC.presence_of_element_located(locator))
+        WDW(self.driver, timer).until(EC.presence_of_element_located(locator))
+        return self.driver.find_element(*locator)
+
+    def _wait_element_present_in_sublement(self, subel, locator: tuple, timer=5):
+        WDW(subel, timer).until(EC.presence_of_element_located(locator))
+        return subel.find_element(*locator)
+
+    def _wait_text_element_in_element(self, element, locator: tuple, timer=5, text=""):
+        WDW(element, timer).until(EC.text_to_be_present_in_element(locator, text))
+        return element.find_element(*locator)
 
     def _set_field_value(self, field: tuple, value):
         field = self._getelement(field)
@@ -104,6 +104,13 @@ class BasePage:
         if element.is_displayed():
             element.clear()
             element.send_keys(value)
+        return element
+
+    def set_element_value_in_input(self, element, value):
+        self._wait_element_present_in_sublement(element, Common.mat_input_element)
+        input_element = element.find_element(*Common.mat_input_element)
+        self.set_element_value(input_element, value)
+        return element
 
     @staticmethod
     def get_checkbox_element_status(element):
@@ -130,6 +137,13 @@ class BasePage:
     def clear_element(self, element):
         element.send_keys(Keys.CONTROL + "a")
         element.send_keys(Keys.BACK_SPACE)
+        return element
+
+    def clear_input_element(self, element):
+        self._wait_element_present_in_sublement(element, Common.mat_input_element)
+        input_element = element.find_element(*Common.mat_input_element)
+        self.clear_element(input_element)
+        return element
 
     def _error_handler(self, error: ERRORS):
         if error in ERRORS.keys():
@@ -520,474 +534,3 @@ class ServiceList(ListPage):
                 service.find_element(Service.config_column).click()
                 return True
         return False
-
-
-class ServiceDetails(Details, ListPage):
-    @property
-    def main_tab(self):
-        self._getelement(Service.main_tab).click()
-        return ListPage(self.driver)
-
-    @property
-    def configuration_tab(self):
-        self._getelement(Service.configuration_tab).click()
-        return Configuration(self.driver)
-
-    @property
-    def status_tab(self):
-        self._getelement(Service.status_tab).click()
-        return ListPage(self.driver)
-
-    def click_main_tab(self):
-        return self._click_tab("Main")
-
-    def click_configuration_tab(self):
-        return self._click_tab("Configuration")
-
-
-# pylint: disable=R0904
-class Configuration(BasePage):
-    """
-    Class for configuration page
-    """
-
-    def assert_field_editable(self, field, editable=True):
-        """Check that we can edit specific field or not
-        :param field:
-        :param editable:
-        :return:
-        """
-        field_editable = self.editable_element(field)
-        assert field_editable == editable
-
-    def assert_field_content_equal(self, field_type, field, expected_value):
-        """Check that field content equal expected value
-
-        :param field_type:
-        :param field:
-        :param expected_value:
-        :return:
-        """
-        current_value = self.get_field_value_by_type(field, field_type)
-        if field_type == 'file':
-            expected_value = 'test'
-        if field_type == 'map':
-            map_config = self.get_map_field_config(field)
-            assert set(map_config.keys()) == set(map_config.keys())
-            assert set(map_config.values()) == set(map_config.values())
-        else:
-            err_message = "Default value wrong. Current value {}".format(current_value)
-            assert current_value == expected_value, err_message
-
-    def assert_alerts_presented(self, field_type):
-        """Check that frontend errors presented on screen and error type in text
-
-        :param field_type:
-        :return:
-        """
-        errors = self.get_frontend_errors()
-        assert errors
-        if field_type == 'password':
-            assert len(errors) == 2
-            error_text = "Field [{}] is required!".format(field_type)
-            error_texts = [error.text for error in errors]
-            assert error_text in error_texts
-
-    def assert_group_status(self, group_element, status=True):
-        """Check that group active or not
-        :param group_element:
-        :param status:
-        :return:
-        """
-        if status:
-            assert self.group_is_active_by_element(group_element)
-        else:
-            assert not self.group_is_active_by_element(group_element)
-
-    def get_config_field(self):
-        return self._getelement(ConfigurationLocators.app_conf_fields)
-
-    def get_app_root_scheme_fields(self, field=None):
-        if not field:
-            return self.driver.find_elements(*ConfigurationLocators.app_root_scheme)
-        else:
-            return field.find_elements(*ConfigurationLocators.app_root_scheme)
-
-    def get_map_key(self, item_element):
-        """Get key value for map field
-
-        :param item_element:
-        :return:
-        """
-        form_field = item_element.find_element(*ConfigurationLocators.map_key_field)
-        inp = form_field.find_element(*Common.mat_input_element)
-        return inp.get_attribute("value")
-
-    def get_map_value(self, item_element):
-        """Get value for map field
-
-        :param item_element:
-        :return:
-        """
-        form_field = item_element.find_element(*ConfigurationLocators.map_value_field)
-        inp = form_field.find_element(*Common.mat_input_element)
-        return inp.get_attribute("value")
-
-    def get_map_field_config(self, map_field):
-        """Get map field values
-
-        :param map_field:
-        :return: dict
-        """
-        items = map_field.find_elements(*Common.item)
-        result = {}
-        for item in items:
-            _key = self.get_map_key(item)
-            _value = self.get_map_value(item)
-            result[_key] = _value
-        return result
-
-    def get_fields_by_type(self, field_type):
-        """Get fields by type
-
-        :param field_type: string with type
-        :return: list of fields
-        """
-        if field_type == 'structure':
-            return self.get_app_root_scheme_fields()
-        return self.get_app_fields()
-
-    def get_field_value_by_type(self, field_element, field_type):
-        """Return field value from element by field type
-
-        :param field_element: WebElement
-        :param field_type: string with type
-        :return: field value, for numeric fields string will be converted
-        to field type.
-        """
-        if field_type == 'boolean':
-            element_with_value = field_element.find_element(*Common.mat_checkbox_class)
-            current_value = self.get_checkbox_element_status(element_with_value)
-        elif field_type == 'option':
-            element_with_value = field_element.find_element(*Common.mat_select)
-            current_value = self.get_field_value(element_with_value)
-        elif field_type == 'list':
-            elements_with_value = field_element.find_elements(*Common.mat_input_element)
-            current_value = [
-                self.get_field_value(element) for element in elements_with_value
-            ]
-        elif field_type == 'structure':
-            return self.get_structure_values(field_element)
-        else:
-            element_with_value = field_element.find_element(*Common.mat_input_element)
-            current_value = self.get_field_value(element_with_value)
-        if field_type == 'integer':
-            current_value = int(current_value)
-        elif field_type == 'float':
-            current_value = float(current_value)
-        elif field_type == 'json':
-            current_value = json.loads(current_value)
-        return current_value
-
-    @staticmethod
-    def get_structure_values(field):
-        """Get structure values for field
-
-        :param field:
-        :return: list of dicts
-        """
-        schemes = field.find_elements(*ConfigurationLocators.app_root_scheme)[1:]
-        config = []
-        for scheme in schemes:
-            fields_in_scheme = scheme.find_elements(*Common.mat_form_field)
-            structure_element = {}
-            for mat_field in fields_in_scheme:
-                input_element = mat_field.find_element(*Common.mat_input_element)
-                name = mat_field.text
-                structure_element[name] = input_element.get_attribute("value")
-            config.append(structure_element)
-        return config
-
-    @staticmethod
-    def get_field_value(input_field):
-        return input_field.get_attribute("value")
-
-    def get_config_elements(self):
-        el = self.get_config_field()
-        return el.find_elements(*Common.all_childs)
-
-    def get_field_groups(self):
-        return self.driver.find_elements(*ConfigurationLocators.field_group)
-
-    def get_group_names(self):
-        return self.driver.find_elements(*ConfigurationLocators.group_title)
-
-    @retry(retry_on_exception=retry_on_exception, stop_max_delay=10 * 1000)
-    def save_button_status(self):
-        self._wait_element_present(ConfigurationLocators.config_save_button)
-        button = self.driver.find_element(*ConfigurationLocators.config_save_button)
-        class_el = button.get_attribute("disabled")
-        if class_el == 'true':
-            result = False
-        else:
-            result = True
-        return result
-
-    def get_app_fields(self):
-        return self.driver.find_elements(*ConfigurationLocators.app_field)
-
-    def _get_group_element_by_name(self, name):
-        self._wait_element_present(Common.mat_expansion_panel)
-        config_groups = self.driver.find_elements(*Common.mat_expansion_panel)
-        for group in config_groups:
-            if name in group.text:
-                return group
-        return False
-
-    def group_is_active_by_element(self, group_element):
-        """
-
-        :param group_element:
-        :return:
-        """
-        toogle = group_element.find_element(*Common.mat_slide_toggle)
-        if 'mat-checked' in toogle.get_attribute("class"):
-            return True
-        return False
-
-    @staticmethod
-    def field_is_ro(field_element):
-        if "read-only" in field_element.get_attribute("class"):
-            return True
-        return False
-
-    def group_is_active_by_name(self, group_name):
-        """Get group status
-        :param group_name:
-        :return:
-        """
-        group = self._get_group_element_by_name(group_name)
-        toogle = group.find_element(*Common.mat_slide_toggle)
-        if 'mat-checked' in toogle.get_attribute("class"):
-            return True
-        return False
-
-    def activate_group_by_element(self, group_element):
-        """Activate group by element
-
-        :param group_element:
-        :return:
-        """
-        toogle = group_element.find_element(*Common.mat_slide_toggle)
-        if 'mat-checked' not in toogle.get_attribute("class"):
-            toogle.click()
-
-    def activate_group_by_name(self, group_name):
-        """
-
-        :param group_name:
-        :return:
-        """
-        group = self._get_group_element_by_name(group_name)
-        toogle = group.find_element(*Common.mat_slide_toggle)
-        if 'mat-checked' not in toogle.get_attribute("class"):
-            toogle.click()
-
-    def save_configuration(self):
-        self._getelement(ConfigurationLocators.config_save_button).click()
-
-    def click_advanced(self):
-        self._wait_element_present(Common.mat_checkbox)
-        buttons = self.driver.find_elements(*Common.mat_checkbox)
-        for button in buttons:
-            if button.get_attribute("textContent").strip() == 'Advanced':
-                self._click_button_element(button)
-                return True
-        return False
-
-    @property
-    def advanced(self):
-        self._wait_element_present(Common.mat_checkbox)
-        buttons = self.driver.find_elements(*Common.mat_checkbox)
-        for button in buttons:
-            if button.get_attribute("textContent").strip() == 'Advanced':
-                return "checked" in button.get_attribute("class")
-        return None
-
-    def show_history(self):
-        for icon in self._getelements(Common.mat_icon):
-            if icon.text == "history":
-                icon.clikc()
-                return True
-        return False
-
-    def _get_tex_boxes_elements(self):
-        textboxes = self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-        return set((textbox.get_attribute("adcm_test"),
-                    textbox.find_element(
-                        *Common.mat_input_element).get_attribute("value")) for textbox in textboxes)
-
-    def _get_text_areas_elements(self):
-        textareas = self._getelements(ConfigurationLocators.app_fields_textarea)
-        areas = []
-        for ar in textareas:
-            el = ar.find_element(Common.textarea)
-            areas.append({ar.get_attribute("adcm_test"): el.get_attribute("value")})
-        return areas
-
-    @staticmethod
-    def _get_tooltip_el_for_field(field):
-        try:
-            return field.find_element(*Common.info_icon)
-        except NoSuchElementException:
-            return False
-
-    @staticmethod
-    def check_tooltip_for_field(field):
-        try:
-            return field.find_element(*Common.info_icon)
-        except NoSuchElementException:
-            return False
-
-    def get_tooltip_text_for_element(self, element):
-        tooltip_icon = self._get_tooltip_el_for_field(element)
-        action = ActionChains(self.driver)
-        action.move_to_element(tooltip_icon).perform()
-        return self.driver.find_element(*Common.tooltip).text
-
-    def _get_map_elements(self):
-        return self._getelements(ConfigurationLocators.app_fields_map)
-
-    def _get_list_elements(self):
-        return self._getelements(ConfigurationLocators.app_fields_list)
-
-    def get_textboxes(self):
-        return self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-
-    def get_labels(self):
-        return self.driver.find_elements(*ConfigurationLocators.app_fields_labels)
-
-    def _get_json_elements(self):
-        json_elements = self._getelements(ConfigurationLocators.app_fields_json)
-        jsons = []
-        for js in json_elements:
-            el = js.find_element(Common.textarea)
-            value = el.get_attribute("value")
-            if value:
-                value = json.loads(value)
-            jsons.append({js.get_attribute("adcm_test"): value})
-        return jsons
-
-    def _get_configuration_elements(self):
-        textboxes = self._getelements(ConfigurationLocators.app_fields_text_boxes)
-        maps = self._getelements(ConfigurationLocators.app_fields_map)
-        passwords = self._getelements(ConfigurationLocators.app_fields_password)
-        textareas = self._getelements(ConfigurationLocators.app_fields_textarea)
-        jsons = self._getelements(ConfigurationLocators.app_fields_json)
-        return textboxes + maps + passwords + textareas + jsons
-
-    def get_password_elements(self):
-        return self.driver.find_elements(*ConfigurationLocators.app_fields_password)
-
-    def _get_config_full_names(self):
-        textboxes = self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-        maps = self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-        passwords = self.driver.find_elements(*ConfigurationLocators.app_fields_password)
-        textareas = self.driver.find_elements(*ConfigurationLocators.app_fields_textarea)
-        jsons = self.driver.find_elements(*ConfigurationLocators.app_fields_json)
-        full_names = []
-        for text_box in textboxes:
-            full_names.append(text_box.get_attribute("adcm_test"))
-        for m in maps:
-            full_names.append(m.get_attribute("adcm_test"))
-        for passw in passwords:
-            full_names.append(passw.get_attribute("adcm_test"))
-        for textarea in textareas:
-            full_names.append(textarea.get_attribute("adcm_test"))
-        for js in jsons:
-            full_names.append(js.get_attribute("adcm_test"))
-        return full_names
-
-    def get_display_names(self):
-        count = 10
-        d_names = {name.text for name in self.driver.find_elements(*Common.display_names)}
-        while count > 0:
-            d_names = {name.text for name in self.driver.find_elements(*Common.display_names)}
-            if not d_names:
-                count -= 1
-                sleep(1)
-            else:
-                break
-        return d_names
-
-    def get_names(self):
-        try:
-            elements = self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-        except StaleElementReferenceException:
-            elements = self.driver.find_elements(*ConfigurationLocators.app_fields_text_boxes)
-        return [textbox.text.split("\n")[0].strip(":") for textbox in elements]
-
-    def set_search_field(self, search_pattern):
-        try:
-            element = self.driver.find_element(*ConfigurationLocators.search_field)
-        except NoSuchElementException:
-            sleep(10)
-            element = self.driver.find_element(*ConfigurationLocators.search_field)
-        self.clear_element(element)
-        self._set_field_value(ConfigurationLocators.search_field, search_pattern)
-
-    def get_group_elements(self):
-        try:
-            return self.driver.find_elements(*Common.display_names)
-        except StaleElementReferenceException:
-            sleep(5)
-            return self.driver.find_elements(*Common.display_names)
-
-    def execute_action(self, action_name):
-        """Click action
-        :param action_name:
-        :return:
-        """
-        assert self._click_button_by_name(action_name, *Common.mat_raised_button)
-        return self._click_button_by_name("Run", *Common.mat_raised_button)
-
-    def element_presented_by_name_and_locator(self, name, by, value):
-        """
-
-        :param name:
-        :param by:
-        :param value:
-        :return:
-        """
-        elements = self.driver.find_elements(by, value)
-        if not elements:
-            return False
-        for el in elements:
-            if el.text == name:
-                return True
-        return False
-
-    @staticmethod
-    def read_only_element(element):
-        """Check that field have read-only attribute
-
-        :param element:
-        :return:
-        """
-        if 'read-only' in element.get_attribute("class"):
-            return True
-        return False
-
-    @staticmethod
-    def editable_element(element):
-        el_class = element.get_attribute("class")
-        el_readonly_attr = element.get_attribute("readonly")
-        if "field-disabled" in el_class:
-            return False
-        elif 'read-only' in el_class:
-            return False
-        elif el_readonly_attr == 'true':
-            return False
-        if element.is_enabled():
-            return True
-        return True
