+++ tests/ui_tests/test_numbers_data/bundles/float-positive/config.yaml
