+++ tests/ui_tests/test_numbers_data/bundles/float-positive_and_null/config.yaml
