+++ tests/ui_tests/test_numbers_data/bundles/integer-negative/config.yaml
