+++ tests/ui_tests/test_numbers_data/bundles/integer-positive_and_negative/config.yaml
