+++ tests/ui_tests/test_password_field.py
@@ -6,8 +6,9 @@ from adcm_client.objects import ADCMClient
 from adcm_pytest_plugin.utils import get_data_dir
 
 from tests.ui_tests.app.app import ADCMTest
+from tests.ui_tests.app.configuration import Configuration
 from tests.ui_tests.app.locators import Common
-from tests.ui_tests.app.pages import Configuration, LoginPage
+from tests.ui_tests.app.pages import LoginPage
 
 
 @pytest.fixture()
